Fixes following debugging and testing
Unfortunately it looks like updating the existing manifest structure
causes a state inconsistency that results in an infinite loop while
looking up segment information. I believe this is just due to the
timeline data being inconsistently represented and this breaks an
invariant of the resolution code.

Digging deeper soon, but the live simulator branch for mpd patches
should provide sufficient test material for anyone that wants to look